[CSL-778] Move base part of Consumer to Lrc.ConsumerB and create Pos.Lrc.Class
